allow time labels at the end of blocks
Enable time labels at the end of a block by inserting no-op
statements that compile to no instructions. Why do I feel a sense
of impending doom like this is inevitably going to be the cause of
dozens of bugs?

Oh well.  As long as we have a good test suite, we'll be fine.
Which we don't yet.  ...I should work on that.