test the VM and fix an AST time bug
Before we can use the VM to test other things, we gotta test the VM
itself!

It requires particularly heavy testing around time labels (and we're
not done!). In testing it so far, I also found a bug in the parser
where apparently the outer statement of a loop was getting its time
set equal to the time at the end of the loop.

*You would think* this could be the cause of #1.  Alas, I recompiled
a new test input .std file using the fixed parser, and it remarkably
still decompiles to the same incorrect result.