cleanup: make visitors easier to use
I've not been a huge fan of publicly exposing visitors because
they're unwieldy to use and you have to remember to call '.finish()'
to check for errors even when they don't return anything.

I tried to hide them earlier, but ran into the issue that many things
that use some of the passes call them on Blocks that are supposed to be
function bodies.  Writing something like an AsFuncBody(Block) was not
feasable as it would have to work for both &Block and &mut Block.

Today I had the idea that Visitable should just treat any block like a
function body.