baby step/cleanup: incorporate type into RawArg
RawArg used to hold the untyped bits of a dword value because the functions
that read binary files had to produce it without any access to type info.

However, now those functions produce something else, and RawArg is now
only a transient thing that briefly exists during raising/lowering, so
there's no reason for it to still be holding untyped data.

Additionally, putting a ScalarValue on this type accomplishes something
that we will need in the very near future: It can now hold a string.