add ResIdent
The modifications to Ident in the previous commit caused some trouble
in places that were expecting Idents to be, well, just dumb idents.
(notably: name resolution itself, and meta objects).  And to be honest,
just generally speaking, I think I feel more comfortable with the name
`Ident` being used for a raw identifier rather than a resolved one.

So this sort of undoes those changes to Ident, and moves them to a
new type which wraps Ident.