resimplify tests, fix final compile-fail bug
This reaps some of the rewards of the prior commit by getting rid
of the instr-related stuff that was recently added to the vm tests
and desugaring tests.

It also tweaks type checking of var declarations to fix the one
remaining broken test.  Huzzah!  Finally, all lights are green in
compile-fail.rs!