use logos lexer
I was considering getting rid of lexer.rs completely and switching
back to the LALRPOP built-in one, but there's the issue of the
virtual tokens.  However, switching to logos at least lets us drop
a lot of the lexing logic.  It's also a tiny bit faster, and
surprisingly speeds up debug builds a couple percent somehow?
(release builds are slightly slower)