change raw register syntax to %REG[10004]
The old `[10004.0]` has always been pretty weird.  There were talks in
Discord earlier about changing it, and I initially didn't feel it was
a big deal. Then yesterday I wanted to add arbitrary expressions to meta,
and ran into an ambiguity between meta arrays and raw registers.

`[10004.0]` is now `%REG[10004]`, and may someday later be `%GVAR_4`.
(with `REG` still available as a fallback for numbers outside of certain
ranges).  The brackets are there to accomodate negative values; I never
was a fan of ideas like `REG_m9999` for `-9999`.