Split ResolveId into ResId and DefId
Name resolution no longer mutates the AST.

The motivation for this is to be able to clone nodes prior to name
resolution, then later be able to perform const evaluation on them
after name resolution. (Without this, the upcoming const solver seemed
kind of intractible.)

To achieve this, ResolveId has been split into two types: DefId, which
uniquely labels a defined "thing", and ResId, which is what gets stored
on idents in the AST and is essentially a "key" for obtaining DefIds
after name resolution.  This design came to mind after looking at the
implementation of rustc.  (ResId is sort of like rustc's NodeId)

I'll attempt to itemize the changes since there's quite a few:

* ResolveId in many places was replaced with DefId, and accordingly,
  many locals named 'res' have been renamed to 'def_id'.  This accounts
  for a lot of unfortunate noise in the commit.
* ResId is quite a bit different from ResolveId, in ways that have
  implications on e.g. the cloning of AST nodes.  See the doc comments
  in this commit.
* A ResId was added to ast::Item::AnmScript.
* There's a new pass to assign ResIds which must be called very soon
  after parsing.

I really wanted to at least have all non-test code compile by the end
of this commit, but unfortunately there's one error that prevents
compilation (name resolution can't insert ScopeEnd statements.)