Remove ResId from AnmScript again
Tests were panicking because the script0 in 'script script0 { ... }'
wasn't resolved.  The detail of defining sequential const integers
for the script names is pretty specific to ANM, so there's no common
code to resolve these.

So just don't make them resolvable, same as the case is in meta with
the sprite names.