cleanup: stop calling it TypeSystem
It's not a type system.  It never was.  It's now called CompilerContext.

There was a smattering of other structs in that module that were at
least vaguely related to types.  ScalarType was moved to crate::value to
sit alongside ScalarType.  InstrAbi and ArgEncoding were moved to
crate::llir.  (both these crate::value and crate::llir moves were
in consideration for a while, so this is a bit of lucky fallout)