combine anm_compile, compile_fail, sbsb, bits_2_bits
For a while, it has been bothering me that `compile-fail.rs` was an
entire file of nothing but compile-fail tests.  Normally, I would
rather be able to juxtapose tests of working code against invalid code,
to make it easier to see the differences between them.

So this takes all of the integration tests that involved compilation or
decompilation of files and combines them into a single "integration_suite"
test runner.  Everything in these files is updated to use the Format
type that compile-fail and sbsb were using, thereby eliminating some
redundant/copypasta code.

Because anm-compile now has access to the same facilities as the other
tests, we no longer need the majority of its resource files; scripts
were inlined into the test files wherever possible.

Tests have been reorganized to put tests next to each other that test
related features, regardless of whether they are compile-fail or not.

(the bits_to_bits tests are still isolated in one file since they have
 no code snippets so there's no value in seeing them next to other
 tests.  Nonetheless, including them in the same test binary reduces
 overall link time of `cargo test`)