add update-dumps.py
I've been reluctant to add this to VCS since it currently has
some absolute paths onto my filesystem, but a friend recently
convinced me that it is better to have at least *something*
in VCS than to have nothing at all.

This is the script I use to perform mass checks of decompilation
and recompilation of all files in all games.