checkpoint: introducing diagnostic thingy
Up until now, errors have always been propagated via `Result` from
from the innermost parts of the compiler, all the way to the very
toplevel code that owns a `Files` object.  And I really enjoyed this,
because basically Spans were the *only thing* that any piece of code
required access to in order to throw an error.

But...... what if we need to generate a warning?  We can't do this with
an `Err`, because the caller of a Result function needs the `Ok` value to
proceed.

Of course, a functional solution is still possible.  I'm pretty sure one
could define a monad that looks like `(Result<T, ()>, Vec<Diagnostic>)`.
But without something like `do` notation we'd be forced to use
`and_then` everywhere and the nested closures would be hell.

So.... I'm not happy about the way in which this increases coupling,
but if we want to be able to generate warnings, we need to pass something
*into* these functions that can print the diagnostics immediately.

This is just an initial commit that's nowhere even close to compiling.