checkpoint: lean WAY back on ErrorOccurred
This reverts a huge deal of the previous commit after I realized that
ErrorOccurred's .already_handled() is basically an instant bug.

I think the better solution is to start treating empty CompileErrors
as legitimate errors, so that they can be returned in cases where all
errors were already emitted.

With this, it is no longer correct to use CompileError as an accumulator.
Instead, a new ErrorStore type has been introduced.