checkpoint: Add lifetime to CompilerContext, but..
So if we want to put DiagnosticHandler on CompilerContext, we need
to give it a lifetime.  However...

    struct Visitor<'a> {
        ctx: &'a mut CompilerContext,
    }

If we try to keep one lifetime parameter, we end up with the classic
problem of an invariant lifetime that borrows from itself:

    struct Visitor<'ctx> {
        ctx: &'ctx mut CompilerContext<'ctx>,
    }

So we actually need two lifetime parameters.

    struct Visitor<'a, 'ctx> {
        ctx: &'a mut CompilerContext<'ctx>,
    }

I really don't like where this is going.  Alternative, please?