checkpoint: Double-lifetime Visitors
Looked a bit harder at rustc and noticed that most visitors actually
have two lifetimes, 'tcx and 'a.  So that's not the reason for their
RefCells.  (seems kind of dumb to think that way in hindsight)

Currently, I don't see much of an advantage to having interior mutability-
at least, not enough to offset the many disadvantages!  I think the only
good thing to come out of it was that it both facilitated (by letting
CompilerContext have a &self -> DefAdder method) and required (in order to
keep RefCell reasoning clean) a design where methods like 'add_local' were
no longer part of CompilerContext.

So I've come back to this commit and have added the missing lifetimes.