add a BinContext to make it slightly more justifiable
Passing a DiagnosticsEmitter around to all of these functions made
me uncomfortable because it was just so...... specific!

BinContext is just a more general sounding thing.  Maybe more things
will be added to it and so it will be passed around to more functions
so that we have less of this situation where a seemingly random set
of functions need access to it.