checkpoint: add BinWriter
I want to start making `.emit()` the standard way to print any
diagnostic, so that we aren't treating errors and warnings so
differently.

This seemed like the perfect time to fix the issue that functions
returning SimpleError have been able to add location info to their
error messages with 'anyhow::Context', but not to their warnings.
We'll exorcise 'anyhow::Error' from the project.

Since this location info is inherently tied to a file, as an
experimental design for how to solve this for binary IO, I made
a BinWriter type to replace BinWrite in common usage, which has
methods for creating diagnostics with location notes instead of spans.