checkpoint: Move methods back to BinRead/BinWrite
In encode_args we want to use these methods on a Cursor<Vec<u8>>.

To finally meet all use cases without overlapping impls, this ended
up needing traits for both the old and new designs of BinRead and
BinWrite.  They reclaim their old role as the trait that provides
most methods used in the read and write functions.

Their "newer" role as trait aliases is now serviced by DynReadSeek
and DynWriteSeek, which now basically solely exist to be used in
the default `dyn DynReadSeek` type args for BinReader.

Annoyingly, this design means that modules in formats/ will need to
import both BinRead AND BinReader.  Ehhhhgh.