checkpoint: fix uses of Encoding
I was *really* overthinking the Encoding stuff for a bit.
A few quick edits and the code now compiles.

Ideally, we'd only have one obvious right way to do things.
So I've been trying to eliminate many places that return CompileError
in favor of emitting diagnostics immediately, to make them similar
to places that emit warnings. (and likewise I've been hoping to get
rid of SimpleError/anyhow::Error entirely)

But here, we want to be able to decode strings both during
binary file reading as well as during script decompilation, and in
former case we want it to use the `loc_diagnostics` stack.  I was
trying to find ways to slip an ErrLocationReporter argument in there
but we don't have one during script decompilation.

So ultimately I kept the signatures of 'Encoded::{decode, encode}'
exactly as is.  For these functions, the caller is responsible for
emitting the error.  While this somewhat muddies the answer to the
question "what is the correct way to produce an error", it is hardly
confusing to see this since returning error payloads in Result::Err
is, well, just idiomatic Rust!

I guess what we need now is a guideline for when we want to emit
immediately, or when we want to return error payloads.