add error::unspanned
Compared to the previous solution of putting methods on BinWriter and
BinReader, this is a more general way of replacing anyhow::Context
that is better applicable to the 'decompile' functions.

    // code that replaces  error::group_anyhow(...).with_context(...);
    // type of loc is &impl unspanned::Annotate
    loc.chain_with(|f| write!(f, "in instruction {}", index), |loc| {
        ...
    })

    // how to emit an error with all labels
    c.emit(error!("{}unknown opcode {}", loc, opcode))

It's certainly not perfect.  Its main usage is inside an error! macro,
where you must supply it at the beginning of the format string without
any separating space (like "{}blah").  There's nothing reminding you to do
this, and nothing preventing you from accidentally adding a space.
Originally I at least had a '.annotate(impl Display) -> impl Display' method
to solve the second point, but using it was deemed far too cumbersome as
it resulted in two nested format-like macro calls:

    c.emit(error!("{}", loc.annotate(format_args!("unknown opcode {}", opcode))))

There's some aspects of it that are not DRY (e.g. the ": " separator must be
repeated in WhileWriting, which couldn't use Node due to type system
limitations).

It's also complicated to understand its proper use due to the reliance on ad-hoc
generic structs, might(?) have a measurable impact on compiletime, and desparately
needs documentation.

......anyhoo, it's going to let us remove SimpleError from the decompile
functions, and that's important enough to offset these disadvantages.