another idea: UnspannedEmitter
I think this design is closer to where I want things to be ultimately.

Like the previous design, it is separate from BinWriter and BinReader
(so it can be used during 'decompile').
The concerns about awkwardly needing to do 'error!("{}message", loc)'
are addressed by making the object have an 'emit' method that takes
'error!("message")', adding the label as a form of postprocessing.

To make it absolutely clear where diagnostics should be emitted in
read/write functions, the emit functions on BinWriter and BinReader were
removed.

There's a couple remaining minor issues, most notably that some errors
in the 'write_*' functions already have spans and so this ends up
printing the filename twice on those diagnostics.  But this can be
easily fixed later by making changes internally within the diagnostics
module, without impacting the interface presented to other modules in
the crate.