checkpoint: pass an &UnspannedEmitter into decompile funcs
Also changed ctx in these functions to be &mut ctx, because that
will be necessary if we ever want to fix #13, and I wanted to make
sure that passing in this &DiagnosticEmitter does not conflict with
that mutable borrow.

Once it is verified that this works, we can easily fix the decompile
methods to stop using SimpleError.

...unfortunately it still doesn't compile right now because Files
actually still uses '&mut self'.  I had mistakenly thought that I
already changed that...