update read_instrs to prepare for MSG fix
This is in preparation for the MSG format change that will have
separate scripts.  This modified design of read_instrs is necessary
to correctly handle scripts that contain an ins_0 at time 0 that isn't
the terminal instruction.

note: This is currently broken for MSG because it hasn't been updated
to supply end_offset yet.  That will be done when updating the format.