use InstrLanguage to resolve raw opcodes
This will be necessary to determine whether some raw 'ins_10'
should use timeline names/signatures or ECL names/signatures.

At first the idea was to track the current language from within
AST visitors and pass it in as arguments during name resolution,
but then each individual pass that wanted to play around with regs
and instructions would have had to repeat the logic for tracking
which language it is currently in. (e.g. switch to Timeline when
entering a timeline item...)

Since the language only affects raw REG and ins_ syntax, I finally
decided to add language fields to these nodes in the AST, and add
an early pass that assigns these.