checkpoint: partial work on defaulting buffer size
The properties truth was calling 'height' and 'width' in ANM are really
runtime buffer dimensions; these are generally the next power of two large
enough to hold the texture.

So really, we want the texture dimensions to be the main properties, and
have default vaules for the buffer properties be derived from those.
The properties have been renamed (with deprecation warnings) to make the
texture properties more attractive to use, and the some of the defaulting
behavior is implemented.
(But not all of it yet because these Options on EntrySpecs really hurt my
brain and it's a shame they have to be there making all of this logic even
more complicated)