refactor source tests
Automatically check for no warnings if expect_warning is not supplied,
and also allow check_compiled to be used together with expect_warning.

This required more sophisticated reasoning about the presence/absence
of various parts of the macro call than could reasonably be achieved
solely through macro repetition syntax, so source_test was reworked into
a fn and the macro just serves as a builder for an input struct.

There is an awkward bit where a closure needs to be taken for giving
check_compile_fail_output! the correct hygiene.  This is made worse by the
fact that said macro already uses this same closure workaround internally,
so we have a macro calling a closure calling a macro calling a closure.
This can probably be simplified, but the two have different enough
responsibilities that I'd rather keep it like this. tch.