Introduce raw integer aliases
I've been avoidant to do this because I wanted it to remain obvious
when a type is still a primitive integer. That issue can be
solved by qualifiying the names under a 'raw::' prefix.

This also addresses the big last TODO that I've been representing
as a panicking unit test.