make the patchouli ins affect all functions in file
This had some annoying implications that really work to show just how
poor our error recovery scheme is.  I put a drop-bomb on a type that
contains fatal errors that could be easily forgotten, and it was very
difficult to fix the rest of the code to stop, well, dropping the bomb!

truth relies too much on '?' to escape from bad situations.
What it SHOULD be doing instead is using more in-band 'Error' values
to recover from errors.  There should also be some sort of method that
allows the program to exit unsuccessfully if and only if at least one
diagnostic of Error severity or higher was printed, so that fewer
functions need to return Result<T, ErrorReported>, and so that the
weird difference between 'return Err(emitter.emit(error!(...)))' for
errors and 'emitter.emit(warning!(...)).ignore();' can become less of
a daily thing.