refactoring: split InstrFormat and LanguageHooks
I wanted to be able to split anm.rs to have a file dedicated to the byte
reading/writing, but some of it is in the InstrFormat impl.

That's when it occurred to me: InstrFormat is not a very accurate name
anymore, since the majority of its functions describe the language rather
than instructions. So this refactoring pulls out a better named "LanguageHooks"
and renames a bunch of locals.

The vast majority of implementors implement both, in part because the
language properties do tend to change with the instruction format
(this is probably why it took me so long to notice the overreach in
responsibilities), and in part because it was the simplest change to
make.  There's one test struct that's slightly less WTF now as it doesn't
implement LanguageHooks, and I'll be splitting up ANM's implementors soon.