make the aliasing error mention EoSD call behavior
I really want to highlight that the main reason why this aliasing check
even exists is mostly to overcome a failure of communication in the
present syntax of truth;  there is no room in the syntax to make it
obvious to the user that function parameters in old ECL use certain
registers, and thus accidental reuse of those registers may occur.

Adding this message helps cement the true purpose of this warning.