modify diff label to affect single statement
difficulty labels that run to the end did not accurately represent how
they're usually used.

In doing this change I had to make difficulty labels no longer
statements; it was too difficult to reason about.  Unfortunately this
will bring back a couple of the pain points that were present back when
time was stored on each instruction... but thankfully not all of them!
(for some examples of why this is less bad than time labels: the parser
is still stateless, and there's no oddities concerning a difficulty label
at the beginning or end of a block)

Furthermore, the syntax has been changed.  'difficulty' was too long
and made things hard to read.  However, I did not want to reserve 'diff'
as a keyword.  (zero also convinced me not to unreserve array-indexing
syntax since we could support user-defined intrinsics for it).

After much deliberation I've decided to go with this abomination.

What used to be:

    difficulty[0x8]:
        stuff;
    difficulty[0xFF]:

Is now:

    {"EN"}: stuff;

Yes, it's a totally out-of-place syntax and we lose the similarities
with 'interrupt'.  Maybe it can become a C-style [[attribute]] later,
I don't know.

Free-standing blocks have also been added to the AST.  These make
it possible to write

    {"EN"}: {
        stmt;
        stmt;
    }

...But also, I've wanted a legitimate excuse to add block statements
for a long time.