stop differentiating RaiseInstr from CompressedInstr
By eliminating the distinction between these, we remove the potentially
costly conversion that might sometimes need to be made just to make
speculative attempts to decompile an instruction a certain way.

Note: I am unable to measure the performance impact of this above
noise.  Either I was concerned over nothing, or the branching in
MaybeDiffSwitch has a comparable cost to allocation.  (think smallvec
performance benchmarks).  But having the MaybeDiffSwitches exist ahead
of time (and having a uniform type for RaiseInstr) helps with cognitive
load and reduces frustration between functions that use different types.