make AbiEncoding not Copy
This will enable a String to be added to AbiEncodings, for enums.

Honestly, I'd rather keep this type as Copy and instead have enums
get assigned Ids, but currently the full construction of an AbiEncoding
is done during parsing, and the enum could be defined in a file that has
not been handled yet. This means a lot of the existing ABI validating logic
would have to be moved to after parsing, and it would take a lot of
effort to Do It Right for a file format that is soon to be deprecated
anyways...