CHECKPOINT: EnumKey
The idea here is to make the various 'automatic consts'
that already exist into just enums, and basically rebrand
all the associated machinery as enum machinery.

In this commit, EnumKey is introduced, but it's not quite
the optimal design yet.  In particular, I failed to realize
early on that, by making this change, the list of the various
types of automatic consts is now the responsibility of Defs,
whereas previously it was part of Raiser.