add . syntax for enum consts
I started to try and write the name resolution logic
for bare ident enum consts and it just felt... wrong.
Too complicated, too delicate, too many questions.
Like what if somebody made an enum const called NAN?

So this adds a syntax with '.' which is the only way to write
an enum const.

I guess this means we're keeping "automatic consts" after all,
and I have to undo a bunch of the "BuiltinEnum" shenanigans.