move EnumKey to llir
EnumKey is still useful for putting in the ABI, but doesn't make much
sense beyond that, so it has been moved out of defs.rs.

I still don't think this is quite what we want though.  These enums
and auto-consts are relevant to type checking, and at some point we'll
eventually want to be able to write them in our own function signatures.
I don't think I have to address that quite yet though.