move EnumKind again, add to signatures
So yeah, uh..... That thing I said we didn't need to do right now?
We need to do it right now.

So this moves the enum back to defs.rs yet again (rebranding it as
TypeColor), but keeps the other changes from the previous commit(s)
which separate how enums and auto consts work.  Maybe I'll undo that
too.  IDFK.  I just REALLY need to write the pass which actually
identifies and substitutes these constants.