Revert "add on-demand const eval"
There's too much stuff relying on it currently.

The thing is, putting const eval calls during lowering makes the
(already complicated!) fallback logic even more complicated. Because
now there end up being weird problems like `20 ? "abc" : "def"` which
need to be const eval'd because of the string type.

Furthermore, the places where we'd need to add it in the stackless
lowerer are not just the places where we currently use .as_const_int()
and friends, but also every time we look up an intrinsic, since e.g.
the function to lower assign-ops directly calls the function to lower
bin-ops.