test proper behavior of $(%RAND)
In PCB ECL, the RAND register behaves differently based on whether
it is read as a float or an int, so its read type must not be optimized
away.

This commit introduces special registers to the AstVm which share some
of the unusual qualities of RAND for testing purposes.  A test is added
to expr_compile.rs which uses one of these registers to verify that
expression compilation does not transform $(%RAND) into $RAND.

Changes are also made to the grammar towards a pending update where
the semantics of `_S` will be divided into `$(...)` and `int(...)`,
and similary for `_f` into `%(...)` and `float(...)`.  This is because
I wanted to use the new syntax in these tests.