steps towards dotless enums
Zero gave this argument regarding dotless enum consts:

> My main concern is that there needs to be a reason for people to use mapfile enums instead of just treating a giant pile of consts like C #defines.
>
> After all, even though enums offer cleaner decompiling and potential type checking, neither of those are immediately obvious to a user and you'd still need to remember/type a . each time.
> Sort of like how some people always use wait instead of setting instruction times purely so that they don't have to remember to structure loops/branches differently just to make it work properly.
>
> So if people start doing like
>     const int Gray = 0;
>     const int Red = 1;
> just to avoid possibly forgetting the period, then that defeats 90% of the point behind having enums to begin with. :ReimuWat:
> That'll be even more relevant if it ever becomes possible to define enums in a script file instead of just a mapfile. An enum declaration and a const pile aren't very different to type, but one would provide type checking at the cost of more typing.
> TLDR, it seems like removing the period would encourage users to do The Proper Thing™ without needing to have the benefits of enum type checking explained to them first.

The implementation will be that these can be immediately mapped to their
definition like any other variable during name resolution, and now the
second pass only applies to those consts which are ambiguous.  The
ambiguous consts will map to a dummy DefId during the first pass,
allowing them to be detected in the second pass.

This commit more or less kills AutoConstKind and replaces them with
named enums, but not all of the syntax changes are here yet.
(nor is the bit of code that acts upon the dummy DefId)