fix bugs in new abi parser
Apparently, when I adopted logos, I stopped updating the `offset` field
(which *used* to track the current location in the file), and didn't
notice that anything was wrong because the only place where it got used
was at the beginning of a file.

This field is now recognized as an "initial offset".