Introduce WorkingAnmFile (broken)
This introduces some differentiation between the struct used to
represent an ANM file while we are importing image sources, versus when
we are reading and writing files.
This is intended to resolve issues where truth was producing errors on
loading textures from image sources that it didn't even plan to use.

Where AnmFile *used* to have a bunch of Options, it is now WorkingAnmFile,
which uses tri-state enums to tell apart explicit versus imported fields.

AnmFile's name has been taken over by a new type, which eliminates most
of the options entirely.

Of note: When an image file is used as an image source, it now simply saves
the path at first, and we do not even read anything or update the entry specs.
It used to be that reading an image file would fill in a default for
`img_width` and `img_height` based on `offset_x` and `offset_y`, but it is
possible for this operation to fail (with offset > image size), and we don't
want it to fail in cases where another source will replace it.

When an ANM file entry is used as an image source, we still copy over all
of the metadata immediately.  (if we did not do this, then it would be tough
to use an ANM file and image directory in tandem, pulling metadata from the
ANM file).  This asymmetry between the behavior of ANM sources and image file
sources is a bit sus, but I think it's justified by the fact that ANM sources
set strictly more fields than directories.

Some problems in this commit:

1. Many tests are broken!!!  I need to do some debugging.

2. Need significantly more tests. (Too bad image source tests are some
of the most time-consuming tests to create...)

3. A design goof slipped in where I put the img_height/img_width fields
on the Texture type.  This is wrong because the original intent of them
not being there was to avoid needing to load images during decompilation.
(see the tex_info field which was added at the last minute...)