fix arg0 logic, make fake file for core mapfiles
Using the new attributes parser was problematic for core mapfiles
because they have no spans.  Rather than making the parser deal with
the possibility of having no spans, I have decided to address this by
writing a mapfile to memory and then parsing it back.

To be honest though, I'm not a fan of having spans from this fake file
showing up in actual error messages, because it is misleading.  I may
want to revert some of this commit's impact.