optional param names, and proper unary intrinsics
Allow param names to be omitted; these params will still be unavailable
for scratch.  Getting the test 'eosd_unnamed_param_still_prevents_scratch'
to underline the unnamed param was tricky, so the assign_registers
function was cleaned up slightly with some of the pieces of mutable state
being given more meaningful names to assist with reasoning.

Fix behavior of unary '-' (which wasn't ever using an UnOp intrinsic) and
allow '~' to fallback to -1-x.

It wasn't intentional for these to both be in the same commit, but here
we are.