move intrinsic definitions to core mapfiles
Default intrinsic mappings ought to be stored in the same place as
default signatures are.  Now they are.

This might make it easier to write AST passes that generate different
things based on the availability of different intrinsics?  Still not
100% sure if I want that to be a thing, but it might be necessitated
by things like the differences in behavior of CountJmp for early ECL.

This commit looks huge but 90% of it is really changing data and not
code.