add shit
The aptly-named "xcrement op" represents pre- and post- increment
and decrement.  It's treated totally separate from other unary operators
as it operates on lvalues, which in truth are basically just vars
and nothing else.

This commit doesn't even let you map these operators to intrinsics;
at this point it does nothing more than extend truth's grammar.
We'll be needing it to deal with CountJmp in some ECL versions where
it is actually equivalent to

    if (x-- > 0) goto label;

There are not currently plans to support the syntax

    x++;

because 'x++' is a value-producing expression.  If anything, then purely
for the sake of consistency, the intrinsic would need to have signature
'SS' and you would be forced to assign the result to a variable.